Integration test changes for readTimeout default change & 2.2.
Some of the connection tests max out stream ids on a connection.  In
protocol v3+ there are up to 2^15 stream ids per connection.  In 2.2,
socketOptions.readTimeout is now set to 12 seconds instead of being
disabled (NODEJS-152).  With 2^15 active requests a read timeout is
likely, so we explicitly disable it for these tests.

There appears to be a slight behavioral change in 2.2 where a host UP
status change event is not sent on a control connection before other
nodes log that the node is up (which triggers CCM to complete the
'start' command).  This behavior is similar to 1.2.x, but 2.0.x and
2.1.x do not exhibit this behavior so removed the 1.2.x exception and
made it apply to all versions.  Will pursue what changed in C* to
cause this (not a bug, but a slight behavior change in C*).