Add service option to DseGssAuthProvider, make it and user optional.
The 'service' parameter is needed in the event that the default value of
'dse' is not sufficient.  In the normal case, it is expected that DSE is
running against the 'dse' user, which is what is used for the service
principal for DSE.  In the event that the user overrides this, they need
to do this here as well.

The 'user' parameter should be optional since it is not needed if there
is only 1 ticket in the krb5 cache.  If the user wants to be explicit,
or if they have multiple tickets in their cache, they can explicitly
provide the user name.  Since 'user' is now optional, it has been moved
after 'host' in the argument list.

Lastly, this changes authGSSClientInit to be called with the service name
instead of the user name, since that is what the kerberos library
expects.