Introduce host pool states
Track states of the HostConnectionPool as a FSM.
Differentiate between pool closing, as a result of change in distance, and shutting down, as a result of client shutdown.
Check the state of the pool when creating it and when increasing its size.
When the control connection host is ignored, move to a different host.