Disable waiting for trace events in parallel query test (#190)
'should query multiple times in parallel with query tracing enabled'
was timing out in CI because of a change introduced in C* 2.2.8/3.0.9
(CASSANDRA-11465) that causes C* to wait for trace events to be
written before returning a response.   For this particular test we
are not interested in the trace responses, which often timeout because
of the amount of burden put on the events table from 2000 queries
that each page several times.