Add required 'localDataCenter' Client option
Adds a new option 'localDataCenter' that is required when using
DCAwareRoundRobinPolicy.

Deprecates the localDc argument in DCAwareRoundRobinPolicy.  If
specified it is still used, even in favor of 'localDataCenter'.

Update tests and examples to use localDataCenter where applicable.

Changes ClientOptions.defaultOptions to not set a default noop
logEmitter.  This was overridding the default Client logEmitter.
Instead, on utils.log if no logEmitter is present (which should
only be the case in tests), noop is used.